+++ reference/portal-cf-worker/.wrangler/tmp/bundle-QI0YHG/middleware-insertion-facade.js
@@ -1,0 +1,11 @@
+				import worker, * as OTHER_EXPORTS from "C:\\Users\\mirko\\OneDrive\\Desktop\\visitportal.dev\\reference\\portal-cf-worker\\src\\worker.ts";
+				import * as __MIDDLEWARE_0__ from "C:\\Users\\mirko\\OneDrive\\Desktop\\visitportal.dev\\node_modules\\.pnpm\\wrangler@3.114.17_@cloudflare+workers-types@4.20260430.1\\node_modules\\wrangler\\templates\\middleware\\middleware-ensure-req-body-drained.ts";
+import * as __MIDDLEWARE_1__ from "C:\\Users\\mirko\\OneDrive\\Desktop\\visitportal.dev\\node_modules\\.pnpm\\wrangler@3.114.17_@cloudflare+workers-types@4.20260430.1\\node_modules\\wrangler\\templates\\middleware\\middleware-miniflare3-json-error.ts";
+
+				export * from "C:\\Users\\mirko\\OneDrive\\Desktop\\visitportal.dev\\reference\\portal-cf-worker\\src\\worker.ts";
+
+				export const __INTERNAL_WRANGLER_MIDDLEWARE__ = [
+					
+					__MIDDLEWARE_0__.default,__MIDDLEWARE_1__.default
+				]
+				export default worker;